tests/requirements.txt
six
pytest
pytest-cache
pytest-cov
pytest-django
pytest-flakes
pytest-pep8
jinja2
pygments
